Create activity_main.xml
This is my first practice project for the Google Android Basics Course.

I wanted to make a card for my son who is at this awkward age where he doesn't like a lot of social situations, but he does love our cats, and one of them in particular. I saw the text for this card "I love cats...it's people I don't like" on a button so decided to use that along with a picture of his favourite cat and a series of doodled icons of annoying people.

Below you see my doodle for the card and the finished card:
 l![card_doodle_nora|374x500](upload://hrGMiQ4q3KiKWpy87GfSAu3vcn.jpg)

![finished_card_nora|259x423](upload://dHdua4t2nRbYaRG912DIwtkzlLU.jpeg)

The design is for the most part similar to the practice birthday card we all did: two differently aligned TextViews and a photo ImageView that fills the background. You can see in the code below, that I introduced some tweaks to the position and styling of the texts: I used different colours, a fatter font (sans-serif). I chose to align the top TextView to the centre of the image, with a bit of extra padding added to the left to stop it from overlapping with the cat's ear. The bottom TextView aligns to the bottom with a larger bottom margin to leave space for the stick figures.

To get the doodled icons right I had to find out about some more tools. I drew the icons myself using [this](https://play.google.com/store/apps/details?id=com.simplemobiletools.draw) app and then edited them in photoshop to make the background transparent [If somebody can tell me how to do this straight in Android Studio, please do!].  

I didn't know how to scale the icons to a specific small size whilst still keeping their proportions. Looking at the [documentation](https://developer.android.com/reference/android/widget/ImageView.ScaleType.html) for ImageView.ScaleType on developer.android.com gave me an inkling that I might have to use android:scaleType="fitXY" but not enough information on how to implement this, so I googled until I found [this answer](https://guides.codepath.com/android/Working-with-the-ImageView) on the site guides.codepath.com. The trick according to these instructions is fix one dimension to the desired dp value, to keep the other one to wrap_content and to use  android:scaleType="fitXY" and the attribute   android:adjustViewBounds="true" to keep the proportions intact, like this:

<ImageView
    android:layout_width="50dp"
    android:layout_height="wrap_content"
    android:scaleType="fitXY"
    android:adjustViewBounds="true"

Once I had scaled my stickmen as desired, I lined them up by giving each of them an ID and then positioning them with the first one aligned to the bottom right corner and each subsequent one left of the previous one. To give them a little breathing space, they each got a margin of 5dp.

The final thing I needed to work out is how to make the stickmen white. I could have done this in photoshop, too, but this time I managed to find a [solution](https://stackoverflow.com/questions/43142062/how-to-change-icon-colors-in-android-studio) how to do it straight in Android Studio by using the attribute android:tint and setting the colour to white, like this:

android:tint="@android:color/white"